Move errors into common
