+++ data/ledger.go
@@ -29,6 +29,7 @@ import (
 	"github.com/algorand/go-algorand/data/committee"
 	"github.com/algorand/go-algorand/data/transactions"
 	"github.com/algorand/go-algorand/ledger"
+	"github.com/algorand/go-algorand/ledger/common"
 	"github.com/algorand/go-algorand/logging"
 	"github.com/algorand/go-algorand/protocol"
 )
@@ -333,7 +334,7 @@ func (l *Ledger) EnsureValidatedBlock(vb *ledger.ValidatedBlock, c agreement.Cer
 		logfn := logging.Base().Errorf
 
 		switch err.(type) {
-		case ledger.BlockInLedgerError:
+		case common.BlockInLedgerError:
 			logfn = logging.Base().Debugf
 		}
 
@@ -361,7 +362,7 @@ func (l *Ledger) EnsureBlock(block *bookkeeping.Block, c agreement.Certificate) 
 				logging.Base().Errorf("unrecoverable protocol error detected at block %d: %v", round, err)
 				protocolErrorLogged = true
 			}
-		case ledger.BlockInLedgerError:
+		case common.BlockInLedgerError:
 			logging.Base().Debugf("could not write block %d to the ledger: %v", round, err)
 			return // this error implies that l.LastRound() >= round
 		default:
