+++ ledger/blockdb.go
@@ -27,6 +27,7 @@ import (
 	"github.com/algorand/go-algorand/agreement"
 	"github.com/algorand/go-algorand/data/basics"
 	"github.com/algorand/go-algorand/data/bookkeeping"
+	"github.com/algorand/go-algorand/ledger/common"
 	"github.com/algorand/go-algorand/logging"
 	"github.com/algorand/go-algorand/protocol"
 )
@@ -89,7 +90,7 @@ func blockGet(tx *sql.Tx, rnd basics.Round) (blk bookkeeping.Block, err error) {
 	err = tx.QueryRow("SELECT blkdata FROM blocks WHERE rnd=?", rnd).Scan(&buf)
 	if err != nil {
 		if err == sql.ErrNoRows {
-			err = ErrNoEntry{Round: rnd}
+			err = common.ErrNoEntry{Round: rnd}
 		}
 
 		return
@@ -104,7 +105,7 @@ func blockGetHdr(tx *sql.Tx, rnd basics.Round) (hdr bookkeeping.BlockHeader, err
 	err = tx.QueryRow("SELECT hdrdata FROM blocks WHERE rnd=?", rnd).Scan(&buf)
 	if err != nil {
 		if err == sql.ErrNoRows {
-			err = ErrNoEntry{Round: rnd}
+			err = common.ErrNoEntry{Round: rnd}
 		}
 
 		return
@@ -118,7 +119,7 @@ func blockGetEncodedCert(tx *sql.Tx, rnd basics.Round) (blk []byte, cert []byte,
 	err = tx.QueryRow("SELECT blkdata, certdata FROM blocks WHERE rnd=?", rnd).Scan(&blk, &cert)
 	if err != nil {
 		if err == sql.ErrNoRows {
-			err = ErrNoEntry{Round: rnd}
+			err = common.ErrNoEntry{Round: rnd}
 		}
 
 		return
@@ -370,12 +371,12 @@ func blockEnsureSingleBlock(tx *sql.Tx) (blk bookkeeping.Block, err error) {
 	err = tx.QueryRow("SELECT MAX(rnd) FROM catchpointblocks").Scan(&max)
 	if err != nil {
 		if err == sql.ErrNoRows {
-			err = ErrNoEntry{}
+			err = common.ErrNoEntry{}
 		}
 		return bookkeeping.Block{}, err
 	}
 	if !max.Valid {
-		return bookkeeping.Block{}, ErrNoEntry{}
+		return bookkeeping.Block{}, common.ErrNoEntry{}
 	}
 	round := basics.Round(max.Int64)
 
@@ -389,7 +390,7 @@ func blockEnsureSingleBlock(tx *sql.Tx) (blk bookkeeping.Block, err error) {
 	err = tx.QueryRow("SELECT blkdata FROM catchpointblocks WHERE rnd=?", round).Scan(&buf)
 	if err != nil {
 		if err == sql.ErrNoRows {
-			err = ErrNoEntry{Round: round}
+			err = common.ErrNoEntry{Round: round}
 		}
 		return
 	}
