+++ ledger/blockqueue.go
@@ -28,6 +28,7 @@ import (
 	"github.com/algorand/go-algorand/agreement"
 	"github.com/algorand/go-algorand/data/basics"
 	"github.com/algorand/go-algorand/data/bookkeeping"
+	"github.com/algorand/go-algorand/ledger/common"
 	"github.com/algorand/go-algorand/logging"
 	"github.com/algorand/go-algorand/protocol"
 	"github.com/algorand/go-algorand/util/metrics"
@@ -195,7 +196,7 @@ func (bq *blockQueue) putBlock(blk bookkeeping.Block, cert agreement.Certificate
 		}
 		bq.mu.Lock()
 
-		return BlockInLedgerError{blk.Round(), nextRound}
+		return common.BlockInLedgerError{blk.Round(), nextRound}
 	}
 
 	if blk.Round() != nextRound {
@@ -219,7 +220,7 @@ func (bq *blockQueue) checkEntry(r basics.Round) (e *blockEntry, lastCommitted b
 	latest = bq.lastCommitted + basics.Round(len(bq.q))
 
 	if r > bq.lastCommitted+basics.Round(len(bq.q)) {
-		return nil, lastCommitted, latest, ErrNoEntry{
+		return nil, lastCommitted, latest, common.ErrNoEntry{
 			Round:     r,
 			Latest:    latest,
 			Committed: lastCommitted,
@@ -236,7 +237,7 @@ func (bq *blockQueue) checkEntry(r basics.Round) (e *blockEntry, lastCommitted b
 func updateErrNoEntry(err error, lastCommitted basics.Round, latest basics.Round) error {
 	if err != nil {
 		switch errt := err.(type) {
-		case ErrNoEntry:
+		case common.ErrNoEntry:
 			errt.Committed = lastCommitted
 			errt.Latest = latest
 			return errt
